datasetgen/requirements.txt
PyYAML==6.0.2
openai==1.62.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pypdf==4.3.1
